More expiry tweaks, new animation mode
1) Simplifies some of the core event loops
2) Animation classes keep track of all greenlets spawned within them so
that expire can gevent.killall them. This works pretty well but should
be investigated further.
3) Use variables instead of integers for inputs (TODO: factor into
separate config)
4) Adds basic “clear” animation that clears all active events on a
strip (TODO: Should it really be called an animation? Maybe this is an
opportunity to create an Action class and subclass an Animation?)
5) Remove try/except blocks where possible (TODO: polish motion tween
animation to not need one)
6) First attempt at implementing increment function and standardizes
vs. the decrement function (TODO: non-uniform, “scaled” decrement.
Increment could probably be optimized too)