Multi-strip support (work-in-progress)
Ok, so it turns out that iterating through the frames on a strip and
putting each one to the client in a for loop isn’t going to work.
Instead we have to concatenate them all together and push them to
channel 0, which was what LEDscape was built for all along. Since I
can’t figure out sysex syncing, this means we can abandon the issue15
branch of LEDscape altogether. This probably makes more sense given the
constraint that you need to define the number of LEDs in each strip as
a constant at the LEDscape config level.

Aside from that, it seems like there is some big time event lag when
you’re spawning a bunch of events on 8 strips. So I guess Kortina and
Russell were right all along in that gevent shouldn’t control timing on
the animations because then we have timing slowing down as there are
more animations. But then how do we control timing better? We could
pre-calculate the state of each frame and return a generator that the
strip can pick frames off of after an elapsed time, but then how do we
have dynamic attributes that can be changed mid-animation?

This commit makes oneshots a list, which probably has better
performance? Doesn’t seem to make a difference for the above but
probably worth putting holds in a list too. Also, this puts strips in a
list as well, because we don’t need channel numbers anymore. Strips now
run their own event queue (separate from the universe). I think this is
probably the right way to do it because otherwise we have to talk to
the universe through the controller via some funky dot logic, which
doesn’t seem right. Use 1 worker to trigger both fire/expire, because
why have a handler for both?  Uh, let’s see. Some other stuff? Idk