Refactored controllers/mappings
First stab at global controller object that animations can grab
parameters from. This still feels a bit janky with the
strip.bind(self.controllers) call, but it seems like the only way to
expose the controllers to the strips to be pass on to the animations.
The via_channel construct is also a bit weird, and probably means we
can only trigger animations from one midi channel at a time. No bueno
:/ Still need to (re-)implement master/pitchwheel, but those should
just act like generic params. The “get” overriding is a bit weird, too.
Idk… this needs work.