Strips now belong to the client
Still not sure if this is the correct approach: to bind the strips to
the client and then iterate through each client, running worker
operations for each strip. We need to figure out how run each strip
worker once and only once (right now, if you want the same set of
strips to be mirrored on two beaglebone’s, those strips’ worker will
still be run separately because they belong to two different clients).
We should figure out how to run each animation once and then
“distribute” that frame to each strip after the fact. But then we run
into the problem of each animation needing to know the explicit length
of each strip it operates on. Perhaps we can use some sort of grouping
algorithm, so that if we’re running the same animation over a bunch of
strips of length “x”, it only runs that animation once and then copies
each frame to those respective strips? Idk…