Pass note range to the animation
Animations should be aware of the possible range of notes for which
they could be called. This is particularly true of the Positional
animation, which determines what position to illuminate an LED for a
given range of midi keys. It looks like the Animation class is getting
a bit fat now. I wonder how we can simplify this…